backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.0
